+++ R/resources.R
@@ -2,7 +2,7 @@
 #'
 #' Lists the names of the Data Resources included in a Data Package.
 #'
-#' @param package List describing a Data Package.
+#' @inheritParams read_resource
 #' @return Character vector with the Data Resource names.
 #' @family read functions
 #' @export